refactor(auth): move `guards`, `decorators` and `constants` to `/common`
